backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Session
python-dotenv
PyMySQL          # ← NEW, SQLAlchemy’s MySQL driver
gunicorn
psycopg2-binary
